requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
pandas>=2.0.0
numpy>=1.24.0
transformers>=4.40.0
datasets>=2.14.0
peft>=0.8.0
trl>=0.8.0
bitsandbytes>=0.41.0
openai>=1.0.0
python-dotenv>=1.0.0
